Functional test for switching dev users
