+++ tests/functional/conftest.py
@@ -1,3 +1,4 @@
+import json
 import os
 import subprocess
 import sys
@@ -93,4 +94,53 @@ def researcher_user(live_server, context):
             },
             "output_checker": False,
         },
+    )
+
+
+@pytest.fixture
+def dev_users(tmp_path, settings):
+    settings.AIRLOCK_DEV_USERS_FILE = tmp_path / "dev_users.json"
+    settings.AIRLOCK_DEV_USERS_FILE.write_text(
+        json.dumps(
+            {
+                "output_checker": {
+                    "token": "output_checker",
+                    "details": {
+                        "username": "output_checker",
+                        "fullname": "Output Checker",
+                        "output_checker": True,
+                        "staff": True,
+                        "workspaces": {},
+                    },
+                },
+                "output_checker_1": {
+                    "token": "output_checker_1",
+                    "details": {
+                        "username": "output_checker_1",
+                        "fullname": "Output Checker 1",
+                        "output_checker": True,
+                        "staff": True,
+                        "workspaces": {},
+                    },
+                },
+                "researcher": {
+                    "token": "researcher",
+                    "details": {
+                        "username": "researcher",
+                        "fullname": "Researcher",
+                        "output_checker": False,
+                        "staff": False,
+                        "workspaces": {
+                            "test-workspace": {
+                                "project_details": {
+                                    "name": "Project 1",
+                                    "ongoing": True,
+                                },
+                                "archived": False,
+                            }
+                        },
+                    },
+                },
+            }
+        )
     )
